VisionTransformers

Documentation for VisionTransformers.

VisionTransformers.ConvAttention

VisionTransformers.MultiHeadAttention

VisionTransformers.MLP

VisionTransformers.img2seq

VisionTransformers.seq2img

VisionTransformers.ConvAttention — Type

ConvAttention(dim::Int; kernel=(3,3), q_stride=(1,1), kv_stride=(1,1), nheads=8, attn_dropout_prob=0.0, proj_dropout_prob=0.0, norm=:BN, qkv_bias=false)
ConvAttention(dim::Int; kernel=(3,3), q_stride=(1,1), kv_stride=(1,1), nheads=8, attn_dropout_prob=0.0, proj_dropout_prob=0.0, qkv_bias=false)

A convolutional attention layer as proposed in CVT.

Parameters

dim: The dimension of the feature embedding.

nheads: The number of heads to use for self attention.

kernel: The kernel size used in the convolutional projection layers.

attn_dropout_prob: Dropout probability in the attention block.

proj_dropout_prob: Dropout probability in the projection block.

q_stride: Convolutional stride used to compute the query.

kv_stride: Convolutional stride used to compute the key and value.

qkv_bias: Whether to include a bias term in the convolutional projection layers.

VisionTransformers.MultiHeadAttention — Type

MultiHeadAttentionModule(qkv, proj; nheads=8, attn_dropout_prob=0.0, proj_dropout_prob=0.0)

VisionTransformers.MLP — Method

MLP(indims, hiddendims, outdims; dropout=0.0, act=Flux.gelu)

Build a 2-layer multi-layer-perceptron.

Parameters

indims: The dimension of the input features.

hiddendims: The dimension of the hidden features.

outdims: The dimension of the output features.

dropout: The dropout probability following each Dense layer.

act: The activation function following the first Dense layer.

VisionTransformers.img2seq — Method

img2seq(x::AbstractArray{<:Any,4})

Take a collection of image tokens of size [W x H x C x N] and flatten them into a sequence of size [C x L x N] where L = W * H.

VisionTransformers.seq2img — Function

seq2img(x::AbstractArray{<:Any,3})

Take a sequence of image tokens of size [C x L x N] and reshape it into an image of size [W x H x C x N], where W = H = sqrt(L).